backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1
redis==5.2.1
prometheus-client==0.21.1
PyJWT==2.10.1
python-json-logger==3.2.1

# Storage layer (wired but optional in MVP)
SQLAlchemy==2.0.37
psycopg[binary]==3.2.4

